# JS- addTask, createElement
I added an event listener to all the buttons that run the addTask function .
1)addTask-
The function recognizes the button and its sibling elements. It takes the content from the input and creates a new task with it at the top of the list.
Then clears the input.

2)createElement-
The function receives all the items needed to build an element (tag type, children, classes, content, etc.), creates a new element and returns it.

No local storage