# JS - changeTaskList + 1 assist function
1)changeTaskList -
The function handles a mouseover event of a specific task. She adds to the document onkeydown and only
During the transition, it checks whether the alt key is pressed in combination with the number 1-3. If so it moves it to the relevant list from 1-3.
After the transfer she will not paint it again until we stand on a new item.

2)taskOnTop-
The function gets a list and an element of a task. It puts the element at the top of the list.

No local storage