# JS -test 4 -  dataReconstruction + 3 assist functions
I added saveNewDataLocal function in the addTask function - so every time we add a task it will update the local storage :
functions:
1) createListElement -
The function receives content and produces a Li-type element with the relevant details and uses the createElement function for this purpose.

2)dataReconstruction -
At the start of running the app it checks if we have existing local storage, if not produces an empty one. If so it will restore the DOM based on the information available in the local storage.

3)saveNewDataLocal-
The function receives a key and an element of a list. It creates a new array containing all the items in the list and updates it in local storage.

4)updateDom-
The function takes the existing information in the local storage and passes it to any key that is considered for the list and updates the relevant list in the DOM (creates new elements based on the information)