#JS - bonus 1: Drag and drop tasks
1) Add event llistener to "dragstart event", and draggabke attribute - To every task

2)draggableDataTransfer -
Create a unique and temporary class and move it with data when dragging a task item

3)Adding Event Listener to each div (section) One cancels default settings of "dragover" event and the other "drop" event to move "task" to div.

4)taskDrop -
The function identifies from the transmitted information the unique class we have assigned and selects the relevant element. It then sends it along with the list element of the current div to the taskOnTop function that passes the "task" to the selected list. It then locally updates the information of both the list from which it was removed and to which a task was added.