# JS -  API fetch : saveToApi & loadFromApi + 3 assist functions - NO error handling and tests not pass
Assist functions:

1)cleanDom-
The function deletes all LI elements from the DOM

2)saveDataLocal-
The function receives complete information, converts it and saves it in the local storage under "tasks"

3)dataReconstruction(change) -
The function now does not restore the DOM but returns the most recent value of JSON-shaped local storage.

4)buttonsHandler -
The function handles all the buttons on the page (since two more have been added) and each of them has a handler assignment.

Main functions:

5)saveToApi-
The function sends a PUT request whose contents are the current memory in the local storage

6)loadFromApi -
The function sends a GET request and extracts the content stored there, then it updates the local storage and the DOM according to what was stored in the API.